Comp+result repo

+++ HorseEvent/Models/Class.cs
@@ -5,7 +5,7 @@
         public int ClassId { get; set; }
         public int DisciplineId { get; set; }
         public string ClassLevel { get; set; }
-        public int? Height { get; set; }
+        public decimal? Height { get; set; }
         public string ProgramName { get; set; }
         public DateTime ClassDate { get; set; }
         public int? CompetitionId { get; set; }
